backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
google-cloud-speech>=2.15.0
websockets>=11.0
anthropic>=0.9.0  # For Claude API access
python-dotenv>=1.0.0  # For environment variable management
